Fix warnings.

+++ Histogram.cpp
@@ -20,7 +20,7 @@ int compare_size_t(const void *va, const void *vb) {
 void printHisogram(size_t *data, size_t sampleSize) {
         qsort(data, sampleSize, sizeof(size_t), compare_size_t);
 
-        int numPercentiles = 15;
+        size_t numPercentiles = 15;
         int percentiles[] = {10,20,30,40,50,60,70,80,90,95,96,97,98,99,100};
 
         size_t max = data[sampleSize - 1];
@@ -31,7 +31,7 @@ void printHisogram(size_t *data, size_t sampleSize) {
                 }
                 size_t num = data[index];
                 std::cout << "\t" << std::setw(3) << percentiles[cur] << "% ";
-                for(int star = 0; star < (num * 60) / max; star++) {
+                for(size_t star = 0; star < (num * 60) / max; star++) {
                         std::cout << "*";
                 }
                 std::cout << std::endl;
